+++ CIM_Framework/openwsman/src/lib/wsman-client.c
@@ -374,14 +374,14 @@ wsmc_options_destroy(client_opt_t * op)
 }
 
 void
-wsmc_set_action_option(client_opt_t * options, unsigned int flag)
+wsmc_set_action_option(client_opt_t * options, unsigned long flag)
 {
-	options->flags |= (unsigned long)flag;
+	options->flags |= flag;
 	return;
 }
 
 
-unsigned int
+unsigned long
 wsmc_get_action_option(client_opt_t * options)
 {
 	return options->flags;
@@ -389,9 +389,9 @@ wsmc_get_action_option(client_opt_t * options)
 
 
 void
-wsmc_clear_action_option(client_opt_t * options, unsigned int flag)
+wsmc_clear_action_option(client_opt_t * options, unsigned long flag)
 {
-	options->flags &= ~((unsigned long)flag);
+	options->flags &= ~flag;
 	return;
 }
 
@@ -648,13 +648,13 @@ wsmc_set_reference(const char *reference, client_opt_t * options)
 }
 
 void
-wsmc_set_sub_expiry(int event_subscription_expire, client_opt_t * options)
+wsmc_set_sub_expiry(float event_subscription_expire, client_opt_t * options)
 {
 	options->expires = event_subscription_expire;
 }
 
 void
-wsmc_set_heartbeat_interval(int heartbeat_interval, client_opt_t * options)
+wsmc_set_heartbeat_interval(float heartbeat_interval, client_opt_t * options)
 {
 	options->heartbeat_interval = heartbeat_interval;
 }
